Rename color utilities package
