+++ packages/color/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "st7789-color",
+  "name": "color",
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -17,7 +17,7 @@
     "dev": "bun run --watch src/index.ts",
     "test": "bun test"
   },
-  "keywords": ["st7789", "color", "tft", "utility", "typescript", "esm"],
+  "keywords": ["color", "utility", "typescript", "esm"],
   "license": "MIT",
   "devDependencies": {
     "@types/bun": "latest",
