+++ packages/st7789-gfx2d/src/demo.ts
@@ -1,6 +1,6 @@
 import { type ST7789, toRGB565, rgba } from "st7789-driver";
 import { Gfx2D } from "./gfx2d.ts";
-import { rgb } from "st7789-color";
+import { rgb } from "color";
 import { DoubleBuffer, TileRenderer } from "./tiles.ts";
